+++ PluginsAndFeatures/azure-toolkit-for-intellij/rider/src/com/microsoft/intellij/runner/webapp/webappconfig/validator/SqlDatabaseValidator.kt
@@ -13,12 +13,6 @@ object SqlDatabaseValidator : ConfigurationValidator() {
 
     private val sqlDatabaseNameRegex = "[\\s]".toRegex()
 
-    private val resourceGroupRegex = "[^\\p{L}0-9-.()_]".toRegex()
-    private const val RESOURCE_GROUP_NAME_MIN_LENGTH = 1
-    private const val RESOURCE_GROUP_NAME_MAX_LENGTH = 90
-    private const val RESOURCE_GROUP_NAME_LENGTH_ERROR =
-            "Resource Group name should be from $RESOURCE_GROUP_NAME_MIN_LENGTH to $RESOURCE_GROUP_NAME_MAX_LENGTH characters"
-
     private val sqlServerNameRegex = "[^a-z0-9-]".toRegex()
     private const val SQL_SERVER_NAME_MIN_LENGTH = 1
     private const val SQL_SERVER_NAME_MAX_LENGTH = 63
@@ -49,7 +43,7 @@ object SqlDatabaseValidator : ConfigurationValidator() {
             validateDatabaseName(model.subscriptionId, model.databaseName, model.sqlServerName)
 
             if (model.isCreatingDbResourceGroup) {
-                validateResourceGroupName(model.subscriptionId, model.dbResourceGroupName)
+                ResourceGroupValidator.validateResourceGroupName(model.subscriptionId, model.dbResourceGroupName)
             } else {
                 checkValueIsSet(model.dbResourceGroupName, UiConstants.RESOURCE_GROUP_NAME_NOT_DEFINED)
             }
@@ -76,33 +70,6 @@ object SqlDatabaseValidator : ConfigurationValidator() {
     @Throws(RuntimeConfigurationError::class)
     private fun validateDatabase(database: SqlDatabase?) {
         if (database == null) throw RuntimeConfigurationError(UiConstants.SQL_DATABASE_NOT_DEFINED)
-    }
-
-    @Throws(RuntimeConfigurationError::class)
-    private fun validateResourceGroupName(subscriptionId: String, name: String) {
-
-        checkValueIsSet(name, UiConstants.RESOURCE_GROUP_NAME_NOT_DEFINED)
-        checkResourceGroupExistence(subscriptionId, name)
-
-        if (name.endsWith('.')) throw RuntimeConfigurationError(UiConstants.RESOURCE_GROUP_NAME_CANNOT_ENDS_WITH_PERIOD)
-
-        validateResourceName(name,
-                RESOURCE_GROUP_NAME_MIN_LENGTH,
-                RESOURCE_GROUP_NAME_MAX_LENGTH,
-                RESOURCE_GROUP_NAME_LENGTH_ERROR,
-                resourceGroupRegex,
-                UiConstants.RESOURCE_GROUP_NAME_INVALID)
-    }
-
-    @Throws(RuntimeConfigurationError::class)
-    private fun checkResourceGroupExistence(subscriptionId: String, resourceGroupName: String) {
-        val subscriptionToResourceGroupMap = AzureModel.getInstance().subscriptionToResourceGroupMap
-
-        if (subscriptionToResourceGroupMap != null) {
-            val resourceGroups = subscriptionToResourceGroupMap.filter { it.key.subscriptionId == subscriptionId }.firstOrNull()?.value
-            if (resourceGroups != null && resourceGroups.any { it.name().equals(resourceGroupName, true) })
-                throw RuntimeConfigurationError(String.format(UiConstants.RESOURCE_GROUP_ALREADY_EXISTS, resourceGroupName))
-        }
     }
 
     /**
